Introduce on demand Goro implementation
The implementation tries to bind to the worker service when one of Goro methods is invoked.
Then it passes all the tasks to Goro instance behind the service and unbinds.

Closes #4.
