+++ .travis.yml
@@ -5,7 +5,7 @@ android:
   - build-tools-22.0.1
   - android-22
   - extra-android-m2repository
-  - sys-img-armeabi-android-10
+  - sys-img-armeabi-v7a-android-15
 
 jdk:
   - oraclejdk8
@@ -14,7 +14,7 @@ install:
   - ./gradlew clean assemble
 
 before_script:
-  - echo no | android create avd --force -n test -t android-10 --abi armeabi
+  - echo no | android create avd --force -n test -t android-15 --abi armeabi-v7a
   - emulator -avd test -no-skin -no-audio -no-window &
   - android-wait-for-emulator
   - adb shell input keyevent 82
