+++ goro/src/test/java/com/stanfy/enroscar/goro/BoundGoroTest.java
@@ -1,20 +1,9 @@
 package com.stanfy.enroscar.goro;
 
-import android.app.Activity;
 import android.content.ComponentName;
-import android.content.Intent;
-import android.os.Build;
-import android.os.IBinder;
+import android.content.ServiceConnection;
 
-import org.junit.Before;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.mockito.InOrder;
-import org.robolectric.Robolectric;
-import org.robolectric.RobolectricGradleTestRunner;
-import org.robolectric.Shadows;
-import org.robolectric.annotation.Config;
-import org.robolectric.shadows.ShadowActivity;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.Executor;
@@ -22,72 +11,40 @@ import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-import static org.assertj.android.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 
 /**
  * Tests for BoundGoro.
  */
-@RunWith(RobolectricGradleTestRunner.class)
-@Config(constants = BuildConfig.class, sdk = Build.VERSION_CODES.LOLLIPOP)
-public class BoundGoroTest {
+public class BoundGoroTest extends BaseBindingGoroTest {
 
   /** Implementation. */
-  private BoundGoro goro;
-
-  /** Mock service instance of Goro. */
-  private Goro serviceInstance;
-
-  private Activity context;
-  private ShadowActivity shadowContext;
-
-  private IBinder binder;
-  private ComponentName serviceCompName;
-  private TestingQueues testingQueues;
+  private BoundGoro.BoundGoroImpl goro;
 
   private BoundGoro.OnUnexpectedDisconnection disconnection;
 
-  @Before
-  public void create() {
-    context = Robolectric.setupActivity(Activity.class);
-    shadowContext = Shadows.shadowOf(context);
+  @Override
+  public void init() {
+    super.init();
     disconnection = mock(BoundGoro.OnUnexpectedDisconnection.class);
-    goro = Goro.bindWith(context, disconnection);
+    goro = (BoundGoro.BoundGoroImpl) Goro.bindWith(context, disconnection);
     goro = spy(goro);
-
-    testingQueues = new TestingQueues();
-    serviceInstance = spy(new Goro.GoroImpl(testingQueues));
-
-    serviceCompName = new ComponentName(context, GoroService.class);
-    GoroService service = new GoroService();
-    binder = new GoroService.GoroBinderImpl(serviceInstance, service.new GoroTasksListener());
-    shadowContext.getShadowApplication()
-        .setComponentNameAndServiceForBindService(
-            serviceCompName,
-            binder
-        );
-    reset(serviceInstance);
   }
 
-  private void assertBinding() {
-    Intent startedService = shadowContext.getNextStartedService();
-    assertThat(startedService).isNotNull();
-    assertThat(startedService).hasComponent(context, GoroService.class);
-    Intent boundService = shadowContext.getNextStartedService();
-    assertThat(boundService).isNotNull();
-    assertThat(boundService).hasComponent(context, GoroService.class);
-
-    verify(goro).onServiceConnected(serviceCompName, binder);
+  @Override
+  protected ServiceConnection serviceConnection() {
+    return goro;
   }
 
-  @Test
-  public void addListenerShouldRecord() {
-    GoroListener listener = mock(GoroListener.class);
-    goro.addTaskListener(listener);
+  @Override
+  protected Goro goro() {
+    return goro;
+  }
+
+  @Override
+  protected void doBinding() {
     goro.bind();
-    assertBinding();
-    verify(serviceInstance).addTaskListener(listener);
   }
 
   @Test
@@ -101,56 +58,12 @@ public class BoundGoroTest {
   }
 
   @Test
-  public void scheduleShouldRecordDefaultQueue() {
-    Callable<?> task = mock(Callable.class);
-    goro.schedule(task);
-    goro.bind();
-    assertBinding();
-    verify(serviceInstance).schedule(Goro.DEFAULT_QUEUE, task);
-  }
-
-  @Test
-  public void scheduleShouldReturnFuture() {
-    Future<?> future = goro.schedule(mock(Callable.class));
-    assertThat(future).isNotNull();
-  }
-
-  @Test
   public void cancelFutureBeforeBindingShouldRemoveRecordedTask() {
     Future<?> future = goro.schedule(mock(Callable.class));
     future.cancel(true);
     goro.bind();
     assertBinding();
     verify(serviceInstance, never()).schedule(anyString(), any(Callable.class));
-  }
-
-  @Test
-  public void scheduleShouldRequestQueueName() {
-    Callable<?> task = mock(Callable.class);
-    goro.schedule("1", task);
-    goro.bind();
-    assertBinding();
-    verify(serviceInstance).schedule("1", task);
-  }
-
-  @Test
-  public void afterBindingAddRemoveListenerShouldBeDelegated() {
-    goro.bind();
-    GoroListener listener = mock(GoroListener.class);
-    goro.addTaskListener(listener);
-    goro.removeTaskListener(listener);
-    InOrder order = inOrder(serviceInstance);
-    order.verify(serviceInstance).addTaskListener(listener);
-    order.verify(serviceInstance).removeTaskListener(listener);
-  }
-
-  @Test
-  public void afterBindingScheduleShouldBeDelegated() {
-    goro.bind();
-    Callable<?> task = mock(Callable.class);
-    ObservableFuture<?> future = mock(ObservableFuture.class);
-    doReturn(future).when(serviceInstance).schedule("2", task);
-    assertThat((Object) goro.schedule("2", task)).isSameAs(future);
   }
 
   @Test
@@ -175,92 +88,6 @@ public class BoundGoroTest {
     verify(task).run();
   }
 
-  @SuppressWarnings("unchecked")
-  @Test
-  public void observersSuccess() throws Exception {
-    Callable<String> task = okTask();
-    FutureObserver<String> observer = mock(FutureObserver.class);
-    goro.schedule(task).subscribe(observer);
-
-    goro.bind();
-    verify(serviceInstance).schedule(Goro.DEFAULT_QUEUE, task);
-    testingQueues.executeAll();
-    verify(observer).onSuccess("ok");
-  }
-
-  @SuppressWarnings("unchecked")
-  private Callable<String> okTask() throws Exception {
-    Callable<String> task = mock(Callable.class);
-    doReturn("ok").when(task).call();
-    return task;
-  }
-
-  @SuppressWarnings("unchecked")
-  @Test
-  public void observersError() throws Exception {
-    Callable<String> task = mock(Callable.class);
-    Exception e = new Exception();
-    doThrow(e).when(task).call();
-    FutureObserver<String> observer = mock(FutureObserver.class);
-    goro.schedule(task).subscribe(observer);
-
-    goro.bind();
-    testingQueues.executeAll();
-    verify(observer).onError(e);
-  }
-
-  @Test
-  public void getOnFuture() throws Exception {
-    final Future<String> f = goro.schedule(okTask());
-    final String[] result = new String[1];
-    final Exception[] error = new Exception[1];
-    Thread t = new Thread() {
-      @Override
-      public void run() {
-        try {
-          result[0] = f.get();
-        } catch (Exception e) {
-          error[0] = e;
-        }
-      }
-    };
-    t.start();
-    assertThat(result).containsNull();
-    goro.bind();
-    testingQueues.executeAll();
-    t.join(1000);
-    if (error[0] != null) {
-      throw new AssertionError(error[0]);
-    }
-    assertThat(result).containsOnly("ok");
-  }
-
-  @Test
-  public void getOnFutureWithTimeout() throws Exception {
-    final Future<String> f = goro.schedule(okTask());
-    final String[] result = new String[1];
-    final Exception[] error = new Exception[1];
-    Thread t = new Thread() {
-      @Override
-      public void run() {
-        try {
-          result[0] = f.get(3, TimeUnit.SECONDS);
-        } catch (Exception e) {
-          error[0] = e;
-        }
-      }
-    };
-    t.start();
-    assertThat(result).containsNull();
-    goro.bind();
-    testingQueues.executeAll();
-    t.join(1000);
-    if (error[0] != null) {
-      throw new AssertionError(error[0]);
-    }
-    assertThat(result).containsOnly("ok");
-  }
-
   @Test
   public void throwingTimeoutWithoutBinding() throws Exception {
     final Future<String> f = goro.schedule(okTask());
@@ -283,24 +110,12 @@ public class BoundGoroTest {
   }
 
   @Test
-  public void clearShouldDelegate() {
+  public void clearShouldNotBePostponedAfterDelegation() {
+    goro = (BoundGoro.BoundGoroImpl) Goro.bindAndAutoReconnectWith(context);
     goro.bind();
-    goro.clear("clearedQueue");
-    verify(serviceInstance).clear("clearedQueue");
-  }
-
-  @Test
-  public void clearShouldBeAppliedAfterBinding() {
-    Callable<?> task1 = mock(Callable.class);
-    goro.schedule("clearedQueue", task1);
-    goro.clear("clearedQueue");
-    Callable<?> task2 = mock(Callable.class);
-    goro.schedule("clearedQueue", task2);
-    goro.bind();
-    InOrder order = inOrder(serviceInstance);
-    order.verify(serviceInstance).schedule("clearedQueue", task1);
-    order.verify(serviceInstance).clear("clearedQueue");
-    order.verify(serviceInstance).schedule("clearedQueue", task2);
+    goro.clear("a");
+    goro.onServiceDisconnected(new ComponentName("any", "any"));
+    verify(serviceInstance).clear("a");
   }
 
   @Test
@@ -313,12 +128,21 @@ public class BoundGoroTest {
 
   @Test
   public void autoReconnection() {
-    goro = spy(Goro.bindAndAutoReconnectWith(context));
+    goro = (BoundGoro.BoundGoroImpl) spy(Goro.bindAndAutoReconnectWith(context));
     goro.bind();
     assertBinding();
     reset(goro);
     goro.onServiceDisconnected(new ComponentName("test", "test"));
     assertBinding();
+  }
+
+  @Test
+  public void afterBindingScheduleShouldBeDelegated() {
+    goro.bind();
+    Callable<?> task = mock(Callable.class);
+    ObservableFuture<?> future = mock(ObservableFuture.class);
+    doReturn(future).when(serviceInstance).schedule("2", task);
+    assertThat((Object) goro.schedule("2", task)).isSameAs(future);
   }
 
 }
